requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.12.4
Faker==9.8.3
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.4.0
psycopg2==2.9.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
whitenoise==5.3.0
